rags/pdf-rag/requirements.txt
chromadb>=0.4.22
PyPDF2>=3.0.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.12.0
python-dotenv>=1.0.0
